CXP-1165: move PeriodEventCountTrait into audit
