+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Audit/Persistence/Traits/PeriodEventCountTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Traits;
+namespace Akeneo\Connectivity\Connection\Infrastructure\Audit\Persistence\Traits;
 
 use Akeneo\Connectivity\Connection\Domain\Audit\Model\AllConnectionCode;
 use Akeneo\Connectivity\Connection\Domain\Audit\Model\Read\HourlyEventCount;
